new feature: add friend

+++ J2018112664江胤佐2/src/main/java/pers/jiangyinzuo/chat/dao/FriendDao.java
@@ -9,5 +9,5 @@ public interface FriendDao {
 	
 	void deleteFriend(Integer userId, Integer friendId);
 	
-	void addFriend(Integer userId, Integer friendId);
+	void addFriend(Long userId, Long friendId);
 }
